chore: move utils into helpers file
